Move fixture up a step in hierarchy
