+++ tests/integration/web/auth/sudo_test.py
@@ -1,10 +1,5 @@
-import pytest
-
-from django.test import RequestFactory
-
 from nav.web.auth.utils import set_account
 from nav.web.auth.sudo import sudo, desudo
-from nav.models.profiles import Account
 
 
 def test_sudo_should_change_session_id(
@@ -33,11 +28,3 @@ def test_desudo_should_change_session_id(
     post_desudo_session_id = session_request.session.session_key
 
     assert pre_desudo_session_id != post_desudo_session_id
-
-
-@pytest.fixture()
-def other_account(db):
-    account = Account(login="other_user")
-    account.save()
-    yield account
-    account.delete()
